+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## Unreleased
 
+## 4.4.0-alpha.4 (2019-04-19)
+* Add options to `website create`: require `--app` or `--org`
+
+## 4.4.0-alpha.3 (2019-04-12)
+* Add options to `website create`: `historyApiRouting`, `indexPage`, `errorPage`
+
+## 4.4.0-alpha.2 (2019-03-05)
+* Add commands: `website publish` and `website unpublish`
+* Adjust existing websites commands to backend changes
+
+## 4.4.0-alpha.1 (2019-02-04)
+* Support websites in organizations
+
+## 4.4.0-alpha.0 (2018-12-14)
+* Add initial websites support
+
 ## 4.1.3 (2019-03-20)
 * Show any 'status' value in all caps (`flex status`)